+++ src/main/java/movie/Movie.java
@@ -1,4 +1,4 @@
-package model;
+package movie;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
